pyproject.toml
[project]
name = "aichatapp"
version = "0.1.0"
description = "AI Chat App"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "ddgs>=9.10.0",
    "dotenv>=0.9.9",
    "fastapi>=0.127.0",
    "gunicorn>=23.0.0",
    "langchain>=1.2.0",
    "langchain-aws>=1.1.1",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "motor>=3.7.1",
    "openai>=2.14.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]


[project.scripts]
start = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
